refactor(prospeo): extract shared parse helpers into utils.ts
