+++ apps/sim/tools/prospeo/search_person.ts
@@ -4,22 +4,8 @@ import {
   type ProspeoSearchPersonParams,
   type ProspeoSearchPersonResponse,
 } from '@/tools/prospeo/types'
+import { parseFiltersObject } from '@/tools/prospeo/utils'
 import type { ToolConfig } from '@/tools/types'
-
-function parseFilters(value: unknown): Record<string, unknown> {
-  if (value && typeof value === 'object' && !Array.isArray(value)) {
-    return value as Record<string, unknown>
-  }
-  if (typeof value === 'string' && value.trim().length > 0) {
-    try {
-      const parsed = JSON.parse(value)
-      if (parsed && typeof parsed === 'object' && !Array.isArray(parsed)) {
-        return parsed as Record<string, unknown>
-      }
-    } catch {}
-  }
-  return {}
-}
 
 export const searchPersonTool: ToolConfig<ProspeoSearchPersonParams, ProspeoSearchPersonResponse> =
   {
@@ -58,7 +44,7 @@ export const searchPersonTool: ToolConfig<ProspeoSearchPersonParams, ProspeoSear
         'Content-Type': 'application/json',
       }),
       body: (params) => {
-        const body: Record<string, unknown> = { filters: parseFilters(params.filters) }
+        const body: Record<string, unknown> = { filters: parseFiltersObject(params.filters) }
         if (params.page !== undefined) body.page = params.page
         return body
       },
